Login now works properly, you now have to enter a valid username and password
Fixed an off-by-one error in PasswordHasherImpl that caused ArrayIndexOutOfBoundsExceptions

The login page now always inserts a 'system' user with password 'aaa' if that is not present, so you can always login with that.
If the database can't be used for some reason you are automatically signed in.
TODO: Remove this behaviour ASAP!